[bitmark-microscope] lots of test/debug code in here
These are the initial testing code while trying to debug the L9935 SPI
stepper motor driver chip.  Just checked in to preserve prior to
deletion ad starting on getting the actual auto-focus running.  This
code will (if 9V battery is used to power Vs) find the home position,
delay then move lens back and forward.  There is a huge delay in the
initialisation to allow this process to run, as timers appear to be
suspended as soon as the primary initialisation is done.

Signed-off-by: Christopher Hall <hsw@bitmark.com>